refactor customerprofile page & moving out tab components
