+++ src/styles/CustomerProfile.css
@@ -1,10 +1,21 @@
+.customers-constraint {
+    width: 100%;
+    height: 100%;
+}
+
+.customers-content {
+    width: 80%;
+    height: 100%;
+    margin-left: 20%;
+}
+
 .title-row .ant-row{
     padding-top: 20px; 
     padding-left: 20px;
 }
 
 .blue-bg {
-    background: #07476b;
+    background: lightslategray;
 }
 
 .bg-height {
